pyproject.toml
[tool.poetry]
name = "sancho"
version = "0.1.0"
description = ""
authors = ["Renan Fonseca <renanfonseca@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
neomodel = "^4.0.5"
tree-sitter = "^0.19.0"
pylint = "^2.10.2"
black = "^21.8b0"
python-dotenv = "^0.19.0"
more-itertools = "^8.8.0"
dgl = "^0.6.1"
torch = "^1.9.0"
pygls = "^0.11.2"
loguru = "^0.5.3"
PyGithub = "^1.55"
GitPython = "^3.1.24"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
lang_server = "sancho.lang_server:main"

[tool.black]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
